// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PlanTier = {
  FREE: 'FREE',
  PRO: 'PRO',
  TEAM: 'TEAM'
} as const

export type PlanTier = (typeof PlanTier)[keyof typeof PlanTier]


export const LedgerEntryType = {
  MONTHLY_GRANT: 'MONTHLY_GRANT',
  TOPUP: 'TOPUP',
  SPEND: 'SPEND',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type LedgerEntryType = (typeof LedgerEntryType)[keyof typeof LedgerEntryType]


export const LedgerOwnerType = {
  USER: 'USER',
  WORKSPACE: 'WORKSPACE'
} as const

export type LedgerOwnerType = (typeof LedgerOwnerType)[keyof typeof LedgerOwnerType]


export const SupabaseBackendStatus = {
  PROVISIONING: 'PROVISIONING',
  READY: 'READY',
  ERROR: 'ERROR'
} as const

export type SupabaseBackendStatus = (typeof SupabaseBackendStatus)[keyof typeof SupabaseBackendStatus]


export const WorkspaceMemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type WorkspaceMemberRole = (typeof WorkspaceMemberRole)[keyof typeof WorkspaceMemberRole]